+++ app/controllers/patients_controller.rb
@@ -9,6 +9,7 @@ class PatientsController < ApplicationController
   def show
     @pipelines = @patient.pipelines.includes(:computations, :patient).
                  order(:iid).order('computations.created_at')
+    @details = Patients::Details.new(@patient.case_number, current_user.token).call
   end
 
   def new